Fix points rendering for Chrome on Mac.
This vertex attribute was used without being enabled.

On almost every platform, it so happened that the attributes here were
being assigned integer values that made the app work correctly. However,
this vertex attribute was actually relying on a vertex attrbute enabled
for one of the other draw calls. In Chrome on Mac, the vertex attributes for
the picking shader are 2,0,1 (or something like that), while they are
2,1,0 on Firefox. This seems like it's probably a property of the GL
driver; maybe Firefox works around it.